Separate device specific code to separate class
It makes more clear what each class does and improves testability. Support for new device types can be added to DeviceDataReader.
